Fixed command line parser and fps calculation.

+++ glwidget.h
@@ -86,7 +86,7 @@ private:
     double currentFps;
     double avgFrames;
     QTimer mTimer;
-    int mFpsTimePerFrame;
+    double mFpsTimePerFrame;
 };
 
 #endif
